Introduce ContractResultAndLogs to share code between TransactionDetails and ContractResultDetails.
