+++ src/pages/TransactionDetails.vue
@@ -215,123 +215,7 @@
       </template>
     </DashboardCard>
 
-    <DashboardCard v-if="contractResult" class="h-card">
-      <template v-slot:title>
-        <span class="h-is-secondary-title">Contract Result</span>
-      </template>
-
-      <template v-slot:leftContent>
-        <Property id="result">
-          <template v-slot:name>Result</template>
-          <template v-slot:value>
-            <StringValue :string-value="contractResult?.result"/>
-          </template>
-        </Property>
-        <Property id="from">
-          <template v-slot:name>From</template>
-          <template v-slot:value>
-            <HexaValue :byte-string="contractResult?.from" v-bind:show-none="true"/>
-          </template>
-        </Property>
-        <Property id="to">
-          <template v-slot:name>To</template>
-          <template v-slot:value>
-            <HexaValue :byte-string="contractResult?.to" v-bind:show-none="true"/>
-          </template>
-        </Property>
-        <Property id="type">
-          <template v-slot:name>Type</template>
-          <template v-slot:value>
-            <StringValue :string-value="contractResult?.type?.toString()"/>
-          </template>
-        </Property>
-        <Property id="errorMessage">
-          <template v-slot:name>Error Message</template>
-          <template v-slot:value>
-            <HexaValue :byte-string ="contractResult?.error_message" v-bind:show-none="true"/>
-          </template>
-        </Property>
-      </template>
-
-      <template v-slot:rightContent>
-        <Property id="gasLimit">
-          <template v-slot:name>Gas Limit</template>
-          <template v-slot:value>
-            <PlainAmount :amount="contractResult?.gas_limit"/>
-          </template>
-        </Property>
-        <Property id="gasUsed">
-          <template v-slot:name>Gas Used</template>
-          <template v-slot:value>
-            <PlainAmount :amount="contractResult?.gas_used"/>
-          </template>
-        </Property>
-        <Property id="maxFeePerGas">
-          <template v-slot:name>Max Fee Per Gas</template>
-          <template v-slot:value>
-            <StringValue :string-value="contractResult?.max_fee_per_gas"/>
-          </template>
-        </Property>
-        <Property id="maxPriorityFeePerGas">
-          <template v-slot:name>Max Priority Fee Per Gas</template>
-          <template v-slot:value>
-            <StringValue :string-value="contractResult?.max_priority_fee_per_gas"/>
-          </template>
-        </Property>
-        <Property id="gasPrice">
-          <template v-slot:name>Gas Price</template>
-          <template v-slot:value>
-            <HbarAmount :amount="contractResult?.gas_price ?? 0" :show-extra="true"/>
-          </template>
-        </Property>
-      </template>
-
-    </DashboardCard>
-
-    <DashboardCard v-if="contractResult?.logs?.length" class="h-card">
-      <template v-slot:title>
-        <span class="h-is-secondary-title">Logs</span>
-      </template>
-
-      <template v-slot:control v-if="contractResult?.logs.length > 2">
-        <div class="is-flex is-justify-content-flex-end is-align-items-baseline">
-          <o-field>
-            <o-select v-model="nbLogLines" class="h-is-text-size-1">
-              <option v-for="n in Math.min(MAX_LOG_LINES, Math.ceil(contractResult?.logs?.length / 2))" :key="n" :value="n">
-                {{ (n === Math.ceil(contractResult?.logs?.length / 2)) ? 'Show all' : 'Show ' + n + (n === 1 ? ' line' :' lines') }}
-              </option>
-            </o-select>
-          </o-field>
-          <button id="prev-block-button" :disabled="logCursor===0"
-                  class="button is-white is-small ml-4" @click="logCursor -= 1">&lt; PREVIOUS
-          </button>
-          <button id="next-block-button" :disabled="logCursor >= contractResult?.logs.length - 2 * nbLogLines"
-                  class="button is-white is-small ml-4" @click="logCursor += 1">NEXT &gt;
-          </button>
-        </div>
-      </template>
-
-      <template v-slot:leftContent>
-
-        <template v-for="l in nbLogLines" :key="l">
-          <ContractResultLog :log="contractResult?.logs[logCursor + l - 1]"/>
-          <hr v-if="l < nbLogLines" class="h-card-separator" style="height: 1px; background: grey"/>
-        </template>
-
-      </template>
-
-      <template v-slot:rightContent>
-
-        <template v-for="l in nbLogLines" :key="l">
-          <ContractResultLog v-if="logCursor + nbLogLines + l - 1 < contractResult?.logs.length"
-                             :log="contractResult?.logs[logCursor + nbLogLines + l - 1]"/>
-          <hr v-if="logCursor + nbLogLines + l - 1 < contractResult?.logs.length - 1 && l < nbLogLines"
-              class="h-card-separator" style="height: 1px; background: grey"/>
-        </template>
-
-      </template>
-
-    </DashboardCard>
+    <ContractResultAndLogs :transaction-id-or-hash="transactionId"/>
 
   </section>
 
@@ -345,7 +229,7 @@
 
 <script lang="ts">
 
-import {computed, defineComponent, inject, onBeforeMount, onMounted, ref, watch} from 'vue';
+import {computed, defineComponent, inject, onBeforeMount, ref, watch} from 'vue';
 import axios, {AxiosResponse} from "axios";
 import {
   AccountBalanceTransactions,
@@ -358,7 +242,6 @@ import {
 import {EntityDescriptor} from "@/utils/EntityDescriptor"
 import {normalizeTransactionId, TransactionID} from "@/utils/TransactionID";
 import {computeNetAmount, makeOperatorAccountLabel, makeTypeLabel} from "@/utils/TransactionTools";
-import {ContractResultDetailsLoader} from "@/components/contract/ContractResultDetailsLoader";
 import {systemContractRegistry} from "@/schemas/SystemContractRegistry";
 import AccountLink from "@/components/values/AccountLink.vue";
 import {base64DecToArr, byteToHex} from "@/utils/B64Utils";
@@ -374,10 +257,9 @@ import Footer from "@/components/Footer.vue";
 import NotificationBanner from "@/components/NotificationBanner.vue";
 import Property from "@/components/Property.vue";
 import DurationValue from "@/components/values/DurationValue.vue";
-import PlainAmount from "@/components/values/PlainAmount.vue";
 import {BlocksResponseCollector} from "@/utils/BlocksResponseCollector";
 import BlockLink from "@/components/values/BlockLink.vue";
-import ContractResultLog from "@/components/contract/ContractResultLog.vue";
+import ContractResultAndLogs from "@/components/transaction/ContractResultAndLogs.vue";
 
 const MAX_INLINE_CHILDREN = 3
 const NB_LOG_LINES = 2
@@ -388,9 +270,8 @@ export default defineComponent({
   name: 'TransactionDetails',
 
   components: {
-    ContractResultLog,
+    ContractResultAndLogs,
     BlockLink,
-    PlainAmount,
     Property,
     NotificationBanner,
     Footer,
@@ -457,16 +338,6 @@ export default defineComponent({
 
     const routeName = computed(() => {
       return entity?.value?.routeName
-    })
-
-    const contractResultDetailsLoader = new ContractResultDetailsLoader(
-        ref(null),
-        ref(null),
-        computed(() => props.transactionId ?? null))
-    onMounted(() => contractResultDetailsLoader.requestLoad())
-    watch(contractResultDetailsLoader.entity, () => {
-      const nbLinesForAll = Math.ceil(contractResultDetailsLoader.entity.value?.logs?.length ?? 0 / 2)
-      nbLogLines.value = Math.min(nbLogLines.value, nbLinesForAll)
     })
 
     onBeforeMount(() => {
@@ -609,7 +480,6 @@ export default defineComponent({
       schedulingTransaction,
       parentTransaction,
       childTransactions,
-      contractResult: contractResultDetailsLoader.entity
     }
   },
 });
@@ -692,14 +562,4 @@ function lookupChildTransactions(transactions: Transaction[]): Transaction[] {
 <!--                                                       STYLE                                                     -->
 <!-- --------------------------------------------------------------------------------------------------------------- -->
 
-<style scoped>
-
-.columns button{
-  vertical-align: initial;
-}
-
-.button.is-small {
-  font-size: 0.65rem;
-}
-
-</style>
+<style/>
